backend/requirements.txt
# Django Core
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Task Scheduling
APScheduler==3.10.4

# Transcription Engine
faster-whisper==1.2.0

# Data Validation
pydantic==2.12.2

# Development Tools
ruff==0.14.1
